move to google fonts and structure template
