+++ packages/template-audiogram/package.json
@@ -16,6 +16,7 @@
     "@remotion/media-utils": "workspace:*",
     "@remotion/captions": "workspace:*",
     "@remotion/zod-types": "workspace:*",
+    "@remotion/google-fonts": "workspace:*",
     "react": "19.0.0",
     "react-dom": "19.0.0",
     "remotion": "workspace:*",
